refactor: userに関するusecaseをディレクトリに格納する
