+++ src/usecases/user/user_repository.go
@@ -1,4 +1,4 @@
-package usecase
+package user
 
 import entity "shortcut_master_api/src/domain"
 
@@ -7,5 +7,6 @@ import entity "shortcut_master_api/src/domain"
 type UserRepository interface {
 	Store(entity.User)
 	Select() []entity.User
+	SelectByEmail(entity.User) entity.User
 	Delete(id string)
 }
